Making it float instead of decimal

+++ reservations_app/views.py
@@ -227,7 +227,7 @@ def accept_reservation(request, id):
         try:
             reservation = Reservation.objects.get(_id=ObjectId(id))
             reservation.status = 'accepted'
-            reservation.prix= Decimal128(reservation.prix)
+            reservation.prix= float(reservation.prix)
             reservation.save()
             return JsonResponse({"message": "Reservation accepted successfully"})
         except Reservation.DoesNotExist:
@@ -241,7 +241,7 @@ def decline_reservation(request, id):
         try:
             reservation = Reservation.objects.get(_id=ObjectId(id))
             reservation.status = 'declined'
-            reservation.prix= Decimal128(reservation.prix)
+            reservation.prix= float(reservation.prix)
             reservation.save()
             return JsonResponse({"message": "Reservation declined successfully"})
         except Reservation.DoesNotExist:
